Modulo Productos Variantes

+++ pazzioli-web-backend/productosmodule/src/main/java/com/pazzioliweb/productosmodule/repositori/BodegasRepository.java
@@ -14,7 +14,7 @@ import com.pazzioliweb.productosmodule.entity.Bodegas;
 import com.pazzioliweb.usuariosbacken.dtos.UsuariobodegasidDTOS;
 
 @Repository
-public interface BodegasRepository extends JpaRepository<Bodegas, Long>{
+public interface BodegasRepository extends JpaRepository<Bodegas, Integer>{
   Optional<Bodegas> findByCodigo(int codigo);
